[FE/#70] Feat : title 내용 새로운 컴포넌트로 분리
